+++ java/src/main/java/codingdojo/overtimecalculator/OvertimeCalculator.java
@@ -4,10 +4,8 @@ import codingdojo.Assignment;
 import codingdojo.Briefing;
 import codingdojo.Overtime;
 
-import java.math.BigDecimal;
-
 public interface OvertimeCalculator {
 
-    Overtime calculate(BigDecimal hours, Assignment assignment, Briefing briefing);
+    Overtime calculate(double hours, Assignment assignment, Briefing briefing);
 
 }
